server/requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
onetimepass==1.0.1
pyserial==3.1.1
six==1.10.0
Werkzeug==0.11.10
wheel==0.26.0
